requirements.txt
discord.py
python-dotenv
aiohttp
matplotlib
numpy
pandas
scikit-learn
statsmodels
prophet
edge-tts
wikipedia
opencv-python